refactor(PDisk): move popup to a separate component
