+++ src/containers/Storage/PDisk/PDisk.scss
@@ -9,10 +9,6 @@
         border-radius: 4px; // to match interactive area with disk shape
     }
 
-    &__popup-wrapper {
-        padding: 12px;
-    }
-
     &__media-type {
         position: absolute;
         top: 0;
