+++ src/containers/Storage/PDisk/PDisk.tsx
@@ -1,21 +1,16 @@
 import React, {useEffect, useState, useRef, useMemo} from 'react';
 import cn from 'bem-cn-lite';
 
-import {Popup} from '@gravity-ui/uikit';
-
-import {InfoViewer} from '../../../components/InfoViewer';
 import {InternalLink} from '../../../components/InternalLink';
 
 import routes, {createHref} from '../../../routes';
-import {EFlag} from '../../../types/api/enums';
 import {TPDiskStateInfo, TPDiskState} from '../../../types/api/pdisk';
-import {getPDiskId} from '../../../utils';
 import {getPDiskType} from '../../../utils/pdisk';
-import {bytesToGB} from '../../../utils/utils';
 
 import {STRUCTURE} from '../../Node/NodePages';
 
 import {DiskStateProgressBar, EDiskStateSeverity} from '../DiskStateProgressBar';
+import {PDiskPopup} from '../PDiskPopup';
 
 import {NOT_AVAILABLE_SEVERITY} from '../utils';
 
@@ -52,7 +47,10 @@ const getStateSeverity = (pDiskState?: TPDiskState) => {
     return isSeverityKey(pDiskState) ? stateSeverity[pDiskState] : NOT_AVAILABLE_SEVERITY;
 };
 
-export const PDisk = ({nodeId, data = {}}: PDiskProps) => {
+export const PDisk = ({nodeId, data: rawData = {}}: PDiskProps) => {
+    // NodeId in data is required for the popup
+    const data = useMemo(() => ({...rawData, NodeId: nodeId}), [rawData, nodeId]);
+
     const [severity, setSeverity] = useState(getStateSeverity(data.State));
     const [isPopupVisible, setIsPopupVisible] = useState(false);
 
@@ -72,45 +70,6 @@ export const PDisk = ({nodeId, data = {}}: PDiskProps) => {
     const hidePopup = () => {
         setIsPopupVisible(false);
     };
-    /* eslint-disable */
-    const preparePdiskData = () => {
-        const {AvailableSize, TotalSize, State, PDiskId, NodeId, Path, Realtime, Device} = data;
-        const errorColors = [EFlag.Orange, EFlag.Red, EFlag.Yellow];
-
-        const pdiskData: {label: string; value: string | number}[] = [
-            {label: 'PDisk', value: getPDiskId({NodeId, PDiskId})},
-        ];
-
-        pdiskData.push({label: 'State', value: State || 'not available'});
-        pdiskData.push({label: 'Type', value: getPDiskType(data) || 'unknown'});
-        NodeId && pdiskData.push({label: 'Node Id', value: NodeId});
-
-        Path && pdiskData.push({label: 'Path', value: Path});
-        pdiskData.push({
-            label: 'Available',
-            value: `${bytesToGB(AvailableSize)} of ${bytesToGB(TotalSize)}`,
-        });
-        Realtime &&
-            errorColors.includes(Realtime) &&
-            pdiskData.push({label: 'Realtime', value: Realtime});
-        Device && errorColors.includes(Device) && pdiskData.push({label: 'Device', value: Device});
-        return pdiskData;
-    };
-    /* eslint-enable */
-
-    const renderPopup = () => (
-        <Popup
-            className={b('popup-wrapper')}
-            anchorRef={anchor}
-            open={isPopupVisible}
-            placement={['top', 'bottom']}
-            // bigger offset for easier switching to neighbour nodes
-            // matches the default offset for popup with arrow out of a sense of beauty
-            offset={[0, 12]}
-        >
-            <InfoViewer title="PDisk" info={preparePdiskData()} size="s" />
-        </Popup>
-    );
 
     const pdiskAllocatedPercent = useMemo(() => {
         const {AvailableSize, TotalSize} = data;
@@ -126,7 +85,7 @@ export const PDisk = ({nodeId, data = {}}: PDiskProps) => {
 
     return (
         <React.Fragment>
-            {renderPopup()}
+            <PDiskPopup data={data} anchorRef={anchor} open={isPopupVisible} />
             <div className={b()} ref={anchor} onMouseEnter={showPopup} onMouseLeave={hidePopup}>
                 <InternalLink
                     to={createHref(
@@ -145,4 +104,4 @@ export const PDisk = ({nodeId, data = {}}: PDiskProps) => {
             </div>
         </React.Fragment>
     );
-}
+};
