+++ src/containers/Storage/PDiskPopup/PDiskPopup.scss
@@ -1,0 +1,3 @@
+.pdisk-storage-popup {
+    padding: 12px;
+}
