+++ src/containers/Storage/PDiskPopup/index.ts
@@ -1,0 +1,1 @@
+export * from './PDiskPopup';
